+++ libevmasm/Ethdebug.cpp
@@ -171,7 +171,13 @@ schema::materials::Compilation materialCompilation(std::vector<Source> const& _s
 
 } // anonymous namespace
 
-Json ethdebug::program(std::string_view _name, unsigned _sourceID, Assembly const& _assembly, LinkerObject const& _linkerObject)
+Json ethdebug::program(
+	std::string_view _name,
+	unsigned _sourceID,
+	Assembly const& _assembly,
+	LinkerObject const& _linkerObject,
+	std::optional<schema::program::Context> _programContext
+)
 {
 	return schema::Program{
 		.compilation = std::nullopt,
@@ -186,7 +192,7 @@ Json ethdebug::program(std::string_view _name, unsigned _sourceID, Assembly cons
 			}
 		},
 		.environment = _assembly.isCreation() ? schema::Program::Environment::CREATE : schema::Program::Environment::CALL,
-		.context = std::nullopt,
+		.context = std::move(_programContext),
 		.instructions = programInstructions(_assembly, _linkerObject, _sourceID)
 	};
 }