+++ libevmasm/Ethdebug.h
@@ -25,6 +25,7 @@
 #include <libevmasm/LinkerObject.h>
 
 #include <map>
+#include <optional>
 
 namespace solidity::evmasm::ethdebug
 {
@@ -41,7 +42,13 @@ struct Source
 };
 
 // returns ethdebug/format/program.
-Json program(std::string_view _name, unsigned _sourceID, Assembly const& _assembly, LinkerObject const& _linkerObject);
+Json program(
+	std::string_view _name,
+	unsigned _sourceID,
+	Assembly const& _assembly,
+	LinkerObject const& _linkerObject,
+	std::optional<schema::program::Context> _programContext = std::nullopt
+);
 
 // returns ethdebug/format/info/resources
 Json resources(